refactor: Move `Carousel`'s modules in Core/UI/Web
