+++ packages/plasma-core/src/components/Carousel/index.ts
@@ -1,7 +1,12 @@
 export { CarouselGridWrapper } from './Carousel';
 export { Carousel } from './Carousel';
-export type { CarouselProps } from './Carousel';
 export { CarouselTrack } from './Carousel';
+
 export { CarouselContext } from './CarouselContext';
+
+export { CarouselItem } from './CarouselItem';
+export type { CarouselItemProps } from './CarouselItem';
+
 export { useCarousel, useCarouselItem, useCarouselContext } from './hooks';
-export type { ScrollAxis } from './types';
+
+export type { CarouselProps, ScrollAxis } from './types';
