+++ packages/plasma-ui/src/components/Carousel/index.ts
@@ -1,10 +1,13 @@
+export { CarouselGridWrapper } from '@sberdevices/plasma-core';
+
+export type { CarouselProps } from '@sberdevices/plasma-core';
+
+export { CarouselItem } from '@sberdevices/plasma-core';
+export type { CarouselItemProps } from '@sberdevices/plasma-core';
+
 export { useCarouselItem, useCarouselContext, useRemoteHandlers } from './Carousel.hooks';
 
-export { Carousel, CarouselGridWrapper } from './Carousel';
-export type { CarouselProps } from './Carousel';
+export { Carousel } from './Carousel';
 
 export { CarouselCol } from './CarouselCol';
 export type { CarouselColProps } from './CarouselCol';
-
-export { CarouselItem } from './CarouselItem';
-export type { CarouselItemProps } from './CarouselItem';
