+++ packages/plasma-web/src/components/Carousel/index.ts
@@ -1,5 +1,5 @@
+export { CarouselItem } from '@sberdevices/plasma-core';
+export type { CarouselItemProps } from '@sberdevices/plasma-core';
+
 export { Carousel, CarouselGridWrapper } from './Carousel';
 export type { CarouselProps } from './Carousel';
-
-export { CarouselItem } from './CarouselItem';
-export type { CarouselItemProps } from './CarouselItem';
